Refactor bindings contract to stable internal module
